# public/swagger.json
{
	"components": {
...
		"schemas": {
			"IResponse": {
				"properties": {
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "boolean"
							}
						]
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ISuccessResponse": {
				"properties": {
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "boolean"
							}
						]
					},
					"message": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"IVerification": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"reset_password_token": {
						"type": "string",
						"nullable": true
					},
					"reset_password_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"verify_email_token": {
						"type": "string",
						"nullable": true
					},
					"verify_email_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"email_verified": {
						"type": "boolean"
					},
					"enable_opt": {
						"type": "boolean"
					},
					"otp_secret": {
						"type": "string",
						"nullable": true
					},
					"otp_verified": {
						"type": "boolean"
					},
					"token_salt": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"reset_password_token",
					"reset_password_token_expires_at",
					"verify_email_token",
					"verify_email_token_expires_at",
					"email_verified",
					"enable_opt",
					"otp_secret",
					"otp_verified",
					"token_salt",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerificationInsert": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"reset_password_token": {
						"type": "string",
						"nullable": true
					},
					"reset_password_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"verify_email_token": {
						"type": "string",
						"nullable": true
					},
					"verify_email_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"email_verified": {
						"type": "boolean"
					},
					"enable_opt": {
						"type": "boolean"
					},
					"otp_secret": {
						"type": "string",
						"nullable": true
					},
					"otp_verified": {
						"type": "boolean"
					},
					"token_salt": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"reset_password_token",
					"reset_password_token_expires_at",
					"verify_email_token",
					"verify_email_token_expires_at",
					"email_verified",
					"enable_opt",
					"otp_secret",
					"otp_verified",
					"token_salt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerificationUpdate": {
				"properties": {
					"id": {
						"type": "string"
					},
					"reset_password_token": {
						"type": "string",
						"nullable": true
					},
					"reset_password_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"verify_email_token": {
						"type": "string",
						"nullable": true
					},
					"verify_email_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"email_verified": {
						"type": "boolean"
					},
					"enable_opt": {
						"type": "boolean"
					},
					"otp_secret": {
						"type": "string",
						"nullable": true
					},
					"otp_verified": {
						"type": "boolean"
					},
					"token_salt": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-authen-ts",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/v1/verification/{id}": {
			"get": {
				"operationId": "GetVerification",
				"responses": {
					"200": {
						"description": "Get verification",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IVerification"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "5",
											"user_id": "test",
											"reset_password_token": "reset_password_token",
											"reset_password_token_expires_at": 1660926192826,
											"verify_email_token": "email_token",
											"verify_email_token_expires_at": 1660926192826,
											"email_verified": false,
											"enable_opt": false,
											"otp_secret": "secret",
											"otp_verified": false,
											"token_salt": "salt"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateVerification",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully update",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerificationUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteVerification",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully delete",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user/verification/{user_id}": {
			"get": {
				"operationId": "GetVerificationByUserId",
				"responses": {
					"200": {
						"description": "Get verification by user_id",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IVerification"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "5",
											"user_id": "test",
											"reset_password_token": "reset_password_token",
											"reset_password_token_expires_at": 1660926192826,
											"verify_email_token": "email_token",
											"verify_email_token_expires_at": 1660926192826,
											"email_verified": false,
											"enable_opt": false,
											"otp_secret": "secret",
											"otp_verified": false,
											"token_salt": "salt"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/verification": {
			"post": {
				"operationId": "AddVerification",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully create",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerificationInsert"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}